// frontend/src/api/generated/boards/boards.ts
export type listBoardsApiV1BoardsGetResponse200 = {
  data: LimitOffsetPageTypeVarCustomizedBoardRead;
  status: 200;
};

export type listBoardsApiV1BoardsGetResponse422 = {
  data: HTTPValidationError;
  status: 422;
};

export type listBoardsApiV1BoardsGetResponseSuccess =
  listBoardsApiV1BoardsGetResponse200 & {
    headers: Headers;
  };
export type listBoardsApiV1BoardsGetResponseError =
  listBoardsApiV1BoardsGetResponse422 & {
    headers: Headers;
  };

export type listBoardsApiV1BoardsGetResponse =
  | listBoardsApiV1BoardsGetResponseSuccess
  | listBoardsApiV1BoardsGetResponseError;